requirements.txt
streamlit==1.31.0
requests==2.31.0
pandas==2.2.0
PyPDF2==3.0.1
python-docx==1.1.0
pyyaml==6.0.1
python-dateutil==2.8.2